move zoneLight and make string search less stupid
